Minor refactor, reorganize file structure, add PlayerFile  class

+++ WorldFilter/world/dimension/region/Location.cs
@@ -35,7 +35,7 @@ namespace WorldFilter {
                 return UnpaddedScaledLength;
             }
             set {
-                UnpaddedScaledLength = (int) value;
+                UnpaddedScaledLength = value;
             }
         }
